move reply button. move reply.js clarification.js to contest.js.
