+++ contest/templates/contest/clarification.html
@@ -17,9 +17,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 -->
-{% load staticfiles %} {% block import_source %}
-<script src="{%static 'contest/js/clarification.js'%}"></script>
-{% endblock import_source %}
+{% load staticfiles %}
 <div class="modal fade" id="clarification" tabindex="-1" role="dialog">
   <div class="modal-dialog">
     <div class="modal-content">
